requirements.txt
# Multi-Strategy Quant Debate Bot Requirements
# =============================================

# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# Trading API
alpaca-py>=0.10.0

# Configuration
python-dotenv>=1.0.0
PyYAML>=6.0

# Web Framework (for existing UI)
Flask>=3.0.0
gunicorn>=21.0.0  # Production WSGI server for cloud deployment

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# LLM Services
google-generativeai>=0.3.0  # Google Gemini API

# Optional: NLP/Sentiment (uncomment if needed)
# transformers>=4.30.0
# torch>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# PDF Generation
weasyprint>=60.0
playwright>=1.40.0  # For reliable PDF export

# Data Storage
pyarrow>=14.0.0  # For parquet file support

# HTML Rendering
Jinja2>=3.1.0
